refactor: extract tunnel module from main.rs
Move ~450 lines of tunnel management code (handle_tunnel_command,
process helpers, cleanup functions) from main.rs into a new
tunnel.rs module. main.rs drops from 625 to 120 lines.

Closes #15
