+++ backend/src/lib.rs
@@ -5,6 +5,7 @@ pub mod handlers;
 pub mod models;
 pub mod scheduler;
 pub mod task_manager;
+pub mod tunnel;
 
 use rust_embed::RustEmbed;
